On way to shader world

+++ include/game.h
@@ -1,6 +1,7 @@
 #ifndef MOONLANDER_GAME_H
 #define MOONLANDER_GAME_H
 
+#include <GL/gl.h>
 #include "window.hpp"
 
 class Game {
@@ -9,7 +10,7 @@ public:
 
     static void initOnceSDL2();
     static void validateSDL2Errors();
-    static void initGL(int width, int height);
+    static void initGL(GLfloat width, GLfloat height);
 };
 
 #endif //MOONLANDER_GAME_H
